✨ Migrate to using `cms` caching, filters, and downloads
The original implementation relied heavily on lots of self-managed
indexes, but `@jitl/notion-api` provides much of this out-of-the-box.
Some speed improvements could be made, but overall, the use of caching
may be useful in the future -- as I'd like to allow for dynamic block
creation to avoid needing to manually assign templates.