✨ Adds the ability to sync BibTeX diffs
Syncing the entire BibTeX exported from Paperpile is usually
counter-productive. Particularly because it results in extraordinarily
long compute times even when changes are small. (Usually, only the
first-time sync of Paperpile is susbstantial.)

To avoid this waste of resources and time (and particularly to propagate
updates as quickly as possible), I've introduced a primitive form of
document diffing to the BibTeXs. (Eseentially, it looks for new keys and
items that differ in their contents.)